resolved integer division ambiguity for the compiler

+++ src/fast_LAPRWM.cpp
@@ -553,9 +553,9 @@ List lap_rwm_C (List initials, List Data, List priors, List scales, List Covs,
     bool adaptCov = as<bool>(control["adaptCov"]);
     // derived from control
     int total_it = n_iter + n_burnin;
-    int its = ceil(total_it / n_block);
-    int n_out = floor(n_iter / n_thin);
-    int start_cov_update = floor(n_burnin / n_block) - 1;
+    int its = ceil((double)total_it/n_block);
+    int n_out = floor((double)n_iter/n_thin);
+    int start_cov_update = floor((double)n_burnin/n_block) - 1;
     uvec keep = seqC(n_burnin + 1, total_it, n_thin);
     // initial scales
     int n_b = b.n_cols;
@@ -943,9 +943,9 @@ List lap_rwm_C_woRE (List initials, List Data, List priors, List scales, List Co
     bool adaptCov = as<bool>(control["adaptCov"]);
     // derived from control
     int total_it = n_iter + n_burnin;
-    int its = ceil(total_it / n_block);
-    int n_out = floor(n_iter / n_thin);
-    int start_cov_update = floor(n_burnin / n_block) - 1;
+    int its = ceil((double)total_it/n_block);
+    int n_out = floor((double)n_iter/n_thin);
+    int start_cov_update = floor((double)n_burnin/n_block) - 1;
     uvec keep = seqC(n_burnin + 1, total_it, n_thin);
     // initial
     int n_Bs_gammas = Bs_gammas.n_rows;
@@ -1136,8 +1136,8 @@ List lap_rwm_C_woRE_nogammas (List initials, List Data, List priors, List scales
     bool adaptCov = as<bool>(control["adaptCov"]);
     // derived from control
     int total_it = n_iter + n_burnin;
-    int its = ceil(total_it / n_block);
-    int n_out = floor(n_iter / n_thin);
+    int its = ceil((double)total_it/n_block);
+    int n_out = floor((double)n_iter/n_thin);
     int start_cov_update = floor(n_burnin / n_block) - 1;
     uvec keep = seqC(n_burnin + 1, total_it, n_thin);
     // initial
